requirements.txt
astroid==2.3.1
colorama==0.4.1
Django==3.0.3
django-appconf==1.0.3
django-ckeditor==5.9.0
django-allauth==0.41.0
django-js-asset==1.2.2
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.2.0
psycopg2==2.8.3
pylint==2.4.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
gunicorn==20.0.4
django-imagekit==4.0.2